feat: Add CreateUserDispenserDto and UpdateUserDispenserDto
These changes add the `CreateUserDispenserDto` and `UpdateUserDispenserDto` classes to the `user-dispenser` module. These DTOs are used for creating and updating user dispensers, respectively. The `CreateUserDispenserDto` extends the `UserDispenser` entity and omits the `active` field, while the `UpdateUserDispenserDto` is a partial type of the `UserDispenser` entity.

chore: Update npm dependencies

This commit updates the npm dependencies to the latest stable versions.

fix: ChangeDispenserDto validation

This commit adds validation to the `ChangeDispenserDto` class in the `user` module. The `user_id` and `dispenser_id` fields are now required and must be numbers.

fix: Update LocalAuthDto example

This commit updates the example value for the `Username` field in the `LocalAuthDto` class in the `auth` module.

chore: Update UserDispenserService

This commit adds the `UserDispenserService` class to the `user-dispenser` module. The service extends the `ModelService` class and includes a `searchFields` property for searching by name.

fix: Update ReferralService

This commit updates the `createReferrals` method in the `ReferralService` class in the `referral` module. The `colours` array has been updated with new color values.

chore: Add UserDispenserModule to CommonModule

This commit adds the `UserDispenserModule` to the `CommonModule` for better organization and module management.

chore: Add UserDispenserService tests

This commit adds tests for the `UserDispenserService` class in the `user-dispenser` module.

chore: Update UserDispenserModule imports

This commit updates the imports in the `UserDispenserModule` to include the necessary modules.

chore: Update CommonModule imports

This commit updates the imports in the `CommonModule` to include the `UserDispenserModule`.

chore: Remove user name from coupon service

This commit removes the user name from the coupon service.