feat: Add CreateLearningQuestionsDto and CreateLearningQuestionsOptionsDto
These changes add the `CreateLearningQuestionsDto` and `CreateLearningQuestionsOptionsDto` classes to the `learning-questions` module. The `CreateLearningQuestionsDto` extends the `LearningQuestions` entity and omits the `active` field. The `CreateLearningQuestionsOptionsDto` extends the `LearningQuestions` entity and includes an `options` field for specifying question options.

chore: Update npm dependencies

This commit updates the npm dependencies to the latest stable versions.

chore: Update LearningQuestionsModule

This commit adds the `LearningQuestionOptionsModule` to the `LearningQuestionsModule` for better organization and module management.

chore: Update LearningQuestionOptionsModule

This commit adds the `LearningQuestionOptionsService` to the `LearningQuestionOptionsModule` for better organization and module management.

chore: Update LearningModule entity

This commit updates the `LearningModule` entity to include a `question_set` field that belongs to the `LearningQuestionSet` entity.

chore: Update LearningQuestionOptions entity

This commit updates the `LearningQuestionOptions` entity to include an `is_correct` field for specifying if an option is correct or not.