refactor: Update CommissionController to include bulk update functionality
This commit updates the `CommissionController` in the `commission.controller.ts` file to include a new endpoint for bulk updating entities. The `updatebulk` method has been added, which accepts a query and an `UpdateCommissionDto` object. It calls the `updateBulkStatus` method in the `CommissionService` to perform the bulk update operation. This change enhances the functionality of the Commission module by allowing bulk updates of entities.