feat: Add certificate image URL property to UserExams entity
This commit adds a new property `certificate_img_url` to the `UserExams` entity in the `user-exams.entity.ts` file. The property is used to store the URL of the certificate image. This change enhances the functionality of the `UserExams` entity by allowing the storage and retrieval of certificate image URLs.