chore: Add learning_completed filter to UserService
The code changes in `user.service.ts` add the `learning_completed` filter to the `UserService` class. This filter is used to retrieve users who have completed their learning process. By including this filter, we ensure that only users who have completed their learning are included in the result set. This change improves the accuracy and effectiveness of the user retrieval process.